PSR-251 Refactoring done proper
